requirements.txt
glfw==2.5.4
imgui==1.4.1
imgui_datascience==0.3.2
keyboard==0.13.5
matplotlib==3.5.3
netifaces==0.11.0
numpy==1.23.2
opencv_python==4.6.0.66
PyOpenGL==3.1.6
pypylon==1.8.0
pyrr==0.10.3
PyVISA==1.12.0
pywin32==304
